+++ config/easy/OrganicCompound/code/CGCode.cpp
@@ -1,0 +1,22 @@
+#include <iostream>
+#include <string>
+#include <vector>
+#include <algorithm>
+
+using namespace std;
+
+/**
+ * PREFIX IN ASCENDING ORDER:
+ * 'meth', 'eth', 'prop', 'but', 'pent', 'hex', 'hept', 'oct', 'non', 'dec'
+ **/
+
+int main()
+{
+    string formula;
+    getline(cin, formula);
+
+    // Write an answer using cout. DON'T FORGET THE "<< endl"
+    // To debug: cerr << "Debug messages..." << endl;
+
+    cout << "organic compound name" << endl;
+}