+++ config/easy/OrganicCompound/code/CGCode.java
@@ -1,0 +1,20 @@
+import java.util.*;
+import java.io.*;
+import java.math.*;
+
+/**
+ * PREFIX IN ASCENDING ORDER:
+ * 'meth', 'eth', 'prop', 'but', 'pent', 'hex', 'hept', 'oct', 'non', 'dec'
+ **/
+class Solution {
+
+    public static void main(String args[]) {
+        Scanner in = new Scanner(System.in);
+        String formula = in.nextLine();
+
+        // Write an answer using System.out.println()
+        // To debug: System.err.println("Debug messages...");
+
+        System.out.println("organic compound name");
+    }
+}