+++ config/easy/OrganicCompound/code/CGCode.js
@@ -1,0 +1,6 @@
+const formula = readline();
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('organic compound name');