+++ config/easy/OrganicCompound/code/CGCode.ts
@@ -1,0 +1,11 @@
+/**
+ * PREFIX IN ASCENDING ORDER:
+ * 'meth', 'eth', 'prop', 'but', 'pent', 'hex', 'hept', 'oct', 'non', 'dec'
+ **/
+
+const formula: string = readline();
+
+// Write an answer using console.log()
+// To debug: console.error('Debug messages...');
+
+console.log('organic compound name');